Remove email-related tests and switch to username login
Removed the checks on profile updates for email field and email verification status. This is a part of an overarching task of replacing email with username as the primary login identifier in our applications. It also entailed changing the login tests from using email to user's name in AuthenticationTest, and altered the registration tests to omit email in RegistrationTest.