Simplified login process and improved UI
Removed the 'update profile information' form from the edit profile view, and the email address field from the login and registration forms, focusing on user's name for the purpose of simplicity and a cleaner user interface. Also, changes were implemented in the navigation layout, converting the logout process into a button action for a better user experience. The link for password recovery was removed from the login view, and replaced with a register prompt to emphasize user registration.