Change authentication from email to username
Authentication was previously done with a user's email. This commit changes the authentication to use the user's name instead. The primary reason for this switch is to simplify the login process, as well as to provide users with a more personalized experience. The changes are reflected in several files, particularly in ConfirmablePasswordController, LoginRequest, User model, RegisteredUserController, and ProfileUpdateRequest by replacing 'email' with 'name'. Please note that unique validation for name has been added where necessary.