Add user authentication check in views
Added checks for user authentication in navigation.blade.php view and meeting.blade.php view. If the user is authenticated, their name will appear, and they'll be allowed to edit meetings. If the user is not authenticated, a Sign-in option will show. This implementation improves user management and restricts non-authenticated users from making unwanted changes but still allows them to view meeting from a link.