Add 'date_id' to fillable property in Vote model
In the Vote model, 'date_id' was added to the 'fillable' property array. This was done to enable mass assignment on 'date_id'. Before this change, only 'voted_by' could be mass assigned but now both 'date_id' and 'voted_by' can be mass assigned. This increases the versatility of the Vote model.