Add validation for empty vote option
A check has been added to validate if any option has been selected for voting in the VoteController. This was necessary to avoid users submitting votes without selecting any options, which could potentially cause errors and incorrect data. Now, users will be redirected back with an error message if no option is selected.