Improve UI feedback and error handling in meeting view
The diff consists multiple improvements in the 'meetings' blade template. It adds a check to display error messages in a more informative way using a notification component if an error occurs during session. Code readability and consistency is improved with better formatting of the view.  Also added a conditional check in views to display 'No dates available for voting' message when 'dates' array is empty, providing a feedback to user in case dates are not available for voting. All checkboxes are unchecked by default. Close button was added.