"Implement voting restriction to prevent multiple votes from the same user"
Added a check that prevents a user from voting on the same meeting multiple times. This is achieved by storing a boolean flag 'voted_meetingId' in the user session after they vote. If the flag exists, they are redirected with an error. On the front end, the voting button is disabled if the user has already voted. The only workaround I found was inspecting element, that changing the disabled button to onclick="votesModal.showModal()", then modal can be opened, but if user tries to save votes this way a check in VoteController.php gives already voted error, and does not save votes.

This solution is not perfect, as users can vote again by clearing cookies or reopening browser. But we assume that meeting creator will only count the votes of names that he knows and will delete other. Writing this also decided to add a message to modal- "It is also recommended to use your real name so that meeting creator knows who voted."