Add validation to VoteController store method
Adding here, because I don't want to create another branch and pr*
Validation checks were added in the store action of VoteController to ensure 'voted_by' field does not exceed 50 characters. Previously the limit was only on the textbox itself, but it was possible to change that limit using inspect element.

pr- pull request