Refactor date creation and deletion flow
Several changes were made mainly to DateController and StoreDate to eliminate unnecessary code and enhance readability. The condition to check if a user is the creator of a meeting is refactored out to `isUserMeetingCreator()` function. This makes the code DRY and readable. In StoreDate, the validation rules were broken down to individual methods making it more structured and interpretable. Fixed wrong date deletion bug. Minor adjustments were made to other files for better type hinting and readability.