Refactor code for readability and maintainability
Several changes were made in this commit to improve code readability and maintainability. Unnecessary comments were removed across various CSS, JS and PHP files. The JavaScript code for the deletion confirmation modal was moved from the 'meeting.blade.php' file to the 'meeting-cards.blade.php' file. This was done because the modal functionality is more related to the meeting cards component and not the individual meeting component. The formatting of form elements in several blade files was also improved to increase legibility.