Adjust branding for beta release
The project branding has been updated to reflect the move to a beta release phase. Specifically, the application name shown in the `navbar` and on the welcome page is now "Timely Beta", and the copyright notice in the footer now includes the version "v1.1.0-beta". This ensures that users are aware they are using a beta version of the application. Also added a footer tto app layout, so all pages now have a footer.