Add export to iCalendar functionality
A new method 'exportToICalendar' has been added to the 'DateController' to allow users to export meeting dates to their iCalendar. This method creates a new route 'export.ics' that returns the iCal file, which can be then directly imported into a user's personal calendar. An 'Export to calendar' button has been added to the UI for ease of use. Currently added this button to all dates, but in the future we will have a separate export button, that will export selected date.