Add web manifest and favicon
In order to improve the UX and the website's presence when saved on user devices, a web manifest was added and favicon links were inserted into the head of the guest, welcome, and app layout pages. These additions ensure that the correct icons are displayed, providing a cohesive branding across all browsers and platforms.