Add Terms of Service and Privacy Policy agreement to register form
An update has been made to the registration form. Users are now asked to confirm their agreement to the Terms of Service and Privacy Policy before registration is allowed. A checkbox and links to the relevant policies have been added. It doesn't link anywhere currently, as both tos and pp are not yet in the site.