Refactor the language selector, remove terms of service link, clean-up tags
The language selector code in both navigation and welcome views is extracted into a new Blade component for cleaner and more maintainable code. The terms of service link is removed from the links in welcome and app layout views as it's no longer required by the website. Lastly, redundant `<li>` tags on privacy policy view and unopened `<body>` tags in welcome view are removed to follow HTML standards.