Refactor meeting-card view to a blade component
Moved meeting-card HTML from dashboard.blade.php to a dedicated blade component for better maintainability and reusability. Also, unnecessary 'showDashboard' function in 'MeetingController.php' was removed because its functionality was replaced with the new blade component. The refactor not only keeps the code DRY but also improves the clarity of the structure.