Refactor MeetingController for better readability and encapsulation
Refactored the logic of methods in the MeetingController by extracting chunks of logic into smaller, specialized private functions. This separates and encapsulates individual functionalities, enhancing its maintainability and readability. Also, for the 'show' method, added checks to ensure 'user' is only fetched when auth is checked, which might improve the performance since we won't be fetching user unnecessarily anymore. Lastly, encapsulated the logic for calculating 'highestVoteCount' into a separate function instead of having it in blade files, adhering to the principle of MVC where views should not contain business logic.