Simplify condition checks and optimize template handling in meeting-cards.blade.php
This commit refactors the condition checks and template handling in meeting-cards.blade.php. To improve code readability, it replaces direct property checking with easier-to-grasp variables, and declutters the HTML structure.

@If statements now use simple variables like $isUserCreator and $is1v1 instead of complex and slow property checking. This enhances performance and conveys the code purpose more clearly. Also, we changed to use variables that encapsulate more complex conditions for easier two-way checks like $highestVotedDates->contains($date->id) and $selectedDate == $date.

Additionally, by separating JavaScript code into a separate file, the HTML template is made cleaner and easier to maintain. This contributes to more maintainable and future-proof code.