Add meeting limit and corresponding messages in controllers
Introduced a limit to the number of meetings a user can create. Once the user reaches the limit of 50 meetings, they are redirected back to the dashboard with an appropriate error message. The error message was also added in the translation files (in both English and Lithuanian) for the MeetingController.

This limit is enforced to prevent spam or misuse of the system by creating a high number of meetings and to ensure that our server resources are not being overstretched.