Revise validation rule for user name length
Reduced the maximum length of the user's name from 255 to 30 in the validation rule in RegisteredUserController. This change was also reflected in the register view by adding the maxlength attribute to the name input field.

This change was made to ensure consistency and prevent potential database constraints, as the User model is set to only accept names with a maximum of 30 characters.